🔒️ Added security for woordenschatOnderdeel + custom requests

+++ api/modules/WoordenschatOnderdeel/WoordenschatOnderdeel.controller.ts
@@ -1,15 +1,20 @@
 import { NextFunction, Response } from "express";
+import ForbiddenError from "../../errors/ForbiddenError";
 import NotFoundError from "../../errors/NotFoundError";
 import { AuthRequest } from "../../middleware/auth/auth.types";
+import { CheckTeacherClasses } from "../../utils";
+import UserService from "../User/User.service";
 import WoordenschatOnderdeelService from "./WoordenschatOnderdeel.service";
 import woordenschatOnderdeelService from "./WoordenschatOnderdeel.service";
 import { WoordenschatOnderdeelBody } from "./WoordenschatOnderdeel.types";
 
 export default class WoordenschatOnderdeelController {
   private woordenschatOnderdeelService: WoordenschatOnderdeelService;
+  private userService: UserService;
 
   constructor() {
     this.woordenschatOnderdeelService = new WoordenschatOnderdeelService();
+    this.userService = new UserService();
   }
 
   all = async (
@@ -17,7 +22,54 @@ export default class WoordenschatOnderdeelController {
     res: Response,
     next: NextFunction
   ) => {
-    const woordenschatOnderdelen = await this.woordenschatOnderdeelService.all({ ...req.body });
+    if (!req.user.isAdmin()) {
+      return new ForbiddenError();
+    }
+
+    const woordenschatOnderdelen = await this.woordenschatOnderdeelService.all({
+      ...req.body,
+    });
+    return res.json(woordenschatOnderdelen);
+  };
+
+  byStudent = async (
+    req: AuthRequest<{ id: number }>,
+    res: Response,
+    next: NextFunction
+  ) => {
+    if (req.user.isStudent()) {
+      req.params.id = req.user.id;
+    }
+
+    if (req.user.isTeacher()) {
+      const klasId = (await this.userService.findOne(req.params.id)).klas.id;
+      if (!(await CheckTeacherClasses(req.user.id, klasId))) {
+        next(new ForbiddenError());
+      }
+    }
+
+    const woordenschatOnderdelen =
+      await this.woordenschatOnderdeelService.byStudent(req.params.id);
+    return res.json(woordenschatOnderdelen);
+  };
+
+  byClass = async (
+    req: AuthRequest<{ id: number }>,
+    res: Response,
+    next: NextFunction
+  ) => {
+    if (req.user.isStudent()) {
+      return new ForbiddenError();
+    }
+
+    if (req.user.isTeacher()) {
+      if (!(await CheckTeacherClasses(req.user.id, req.params.id))) {
+        next(new ForbiddenError());
+      }
+    }
+
+    const woordenschatOnderdelen =
+      await this.woordenschatOnderdeelService.byClass(req.params.id);
     return res.json(woordenschatOnderdelen);
   };
 
@@ -26,7 +78,8 @@ export default class WoordenschatOnderdeelController {
     res: Response,
     next: NextFunction
   ) => {
-    const woordenschatOnderdeel = await this.woordenschatOnderdeelService.findOne(req.params.id);
+    const woordenschatOnderdeel =
+      await this.woordenschatOnderdeelService.findOne(req.params.id);
 
     if (!woordenschatOnderdeel) {
       next(new NotFoundError());
@@ -39,9 +92,14 @@ export default class WoordenschatOnderdeelController {
     res: Response,
     next: NextFunction
   ) => {
+    if (req.user.isTeacher()) {
+      return new ForbiddenError();
+    }
+
     const { body } = req;
 
-    const woordenschatOnderdeel = await this.woordenschatOnderdeelService.create(body);
+    const woordenschatOnderdeel =
+      await this.woordenschatOnderdeelService.create(body);
 
     return res.json(woordenschatOnderdeel);
   };
@@ -51,14 +109,31 @@ export default class WoordenschatOnderdeelController {
     res: Response,
     next: NextFunction
   ) => {
-    const { body } = req;
+    let { body } = req;
     body.id = parseInt(req.params.id);
 
+    if (req.user.isTeacher()) {
+      const woordenschatOnderdeel =
+        await this.woordenschatOnderdeelService.findOne(
+          parseInt(req.params.id)
+        );
+      body = { ...woordenschatOnderdeel, feedback: body.feedback };
+    }
+
+    if (req.user.isStudent()) {
+      const woordenschatOnderdeel =
+        await this.woordenschatOnderdeelService.findOne(
+          parseInt(req.params.id)
+        );
+      body.feedback = woordenschatOnderdeel.feedback;
+    }
+
     try {
-      const woordenschatOnderdeel = await this.woordenschatOnderdeelService.update(
-        parseInt(req.params.id),
-        req.body
-      );
+      const woordenschatOnderdeel =
+        await this.woordenschatOnderdeelService.update(
+          parseInt(req.params.id),
+          req.body
+        );
       if (!woordenschatOnderdeel) {
         next(new NotFoundError());
       }
@@ -73,8 +148,13 @@ export default class WoordenschatOnderdeelController {
     res: Response,
     next: NextFunction
   ) => {
+    if (req.user.isTeacher()) {
+      return new ForbiddenError();
+    }
+
     try {
-      const woordenschatOnderdeel = await this.woordenschatOnderdeelService.delete(parseInt(req.params.id));
+      const woordenschatOnderdeel =
+        await this.woordenschatOnderdeelService.delete(parseInt(req.params.id));
       if (!woordenschatOnderdeel) {
         next(new NotFoundError());
       }
